Rename bot commands to register commands
