+++ bot/pint_bot.py
@@ -5,14 +5,13 @@ import time
 import discord
 from discord.ext import commands
 from dotenv import load_dotenv
-from bot.bot_commands import register_commands
 import bot.config as config
+from bot.register_commands import register_commands
 
 intents = discord.Intents.default()
 intents.message_content = True
 
 bot = commands.Bot("!", intents=intents)
-
 
 @bot.event
 async def on_ready():
