+++ bot/register_commands.py
